// job-portal-fe-admin/src/model/jobPositionDetailDto.ts
/**
 * job-portal
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: local
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import {ReferenceDto} from './referenceDto';
import {ReferenceUserDto} from "./referenceUserDto";

export interface JobPositionDetailDto {
  applied?: boolean;
  appliedCount?: number;
  city?: string;
  company?: ReferenceUserDto;
  contactEmail?: string;
  country?: string;
  created?: Date;
  detail?: string;
  favourite?: boolean;
  id?: string;
  jobCategories?: Array<ReferenceDto>;
  lastUpdated?: Date;
  positionName?: string;
  state?: string;
  status?: JobPositionDetailDto.StatusEnum;
}

export namespace JobPositionDetailDto {
  export type StatusEnum = 'ACTIVE' | 'INACTIVE';
  export const StatusEnum = {
    ACTIVE: 'ACTIVE' as StatusEnum,
    INACTIVE: 'INACTIVE' as StatusEnum
  };
}