// job-portal-fe-admin/src/model/jobPositionUpdateDto.ts
/**
 * job-portal
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: local
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface JobPositionUpdateDto { 
    city?: string;
    contactEmail?: string;
    country?: string;
    detail?: string;
    jobCategories?: Array<string>;
    positionName?: string;
    state?: string;
    status: JobPositionUpdateDto.StatusEnum;
}
export namespace JobPositionUpdateDto {
    export type StatusEnum = 'ACTIVE' | 'INACTIVE';
    export const StatusEnum = {
        ACTIVE: 'ACTIVE' as StatusEnum,
        INACTIVE: 'INACTIVE' as StatusEnum
    };
}